part of v3
to make these work on the server, change the SERVER variable to the IP address (it's also in the comment below that line)

v1 is the first versions of the scripts that were made just for testing.
v2 are the newer nicer function scripts that will serve as the basis for the game's connection scripts.
v3 (coming soon) will be made using v2's scripts but with more defined roles, including ServerManager and GameServer.

still working on v3, but here's my progress so far. a new game is set on a new port. gotta make the part where the player connects/switches to that port.